+++ src/dto/userDTO.js
@@ -4,6 +4,7 @@ class UserDto {
       this.user = user.user;
       this.email = user.email;
       this.role = role;
+      this.cart = cart;
     }
 }